Implement all observable/operators using the new useImperativeHandle structure
