+++ app/components/track/parts/Ramp.tsx
@@ -1,15 +1,24 @@
-import { BallDetector } from "../../BallDetector";
-import { Element } from "../elements/Element";
+import type { ForwardedRef } from "react";
+import { forwardRef } from "react";
+import { BuildTail } from "~/components/Build";
+import type { OperatorBuilder } from "~/types";
+import type { Part, TrackPart } from "../parts";
+import { Ramp as RampPart } from "../../elements/Ramp";
 
-type Props = JSX.IntrinsicElements["group"];
+type Props = {
+  track: TrackPart<Part.Ramp>;
+};
 
-export function Ramp({ ...props }: Props) {
+export const Ramp = forwardRef(function Ramp(
+  { track }: Props,
+  ref: ForwardedRef<OperatorBuilder>
+) {
   return (
-    <group {...props}>
-      <group position={[1, 0, 0]}>
-        <Element name="Cube034" rotation={[0, Math.PI / 2, 0]} />
-        <BallDetector position={[0, 1, 0]} />
+    <group>
+      <RampPart />
+      <group position={[2, -1, 0]}>
+        <BuildTail ref={ref} track={track.tail} />
       </group>
     </group>
   );
-}
+});
